+++ requests_cache/models/__init__.py
@@ -7,7 +7,7 @@ from requests import PreparedRequest, Request, Response
 from .base import RichMixin
 from .raw_response import CachedHTTPResponse
 from .request import CachedRequest
-from .response import CachedResponse, OriginalResponse
+from .response import CachedResponse, DecodedContent, OriginalResponse
 
 AnyResponse = Union[OriginalResponse, CachedResponse]
 AnyRequest = Union[Request, PreparedRequest, CachedRequest]