+++ requests_cache/serializers/__init__.py
@@ -1,4 +1,22 @@
 """Response serialization utilities. See :ref:`serializers` for general usage info.
+
+**Summary:**
+
+The ``cattrs`` library includes a number of `pre-configured converters
+<https://cattrs.readthedocs.io/en/latest/preconf.html>`_ that perform some pre-serialization steps
+required for specific serialization formats.
+
+The module :py:mod:`requests_cache.serializers.preconf` wraps those converters as serializer
+:py:class:`.Stage` objects, which are then combined into a :py:class:`.SerializerPipeline`. Preconf
+converters run after the base converter and before the format's ``dumps()`` (or equivalent) method.
+For example, for JSON:
+
+* Run base converter (:py:class:`.CattrStage`) to convert :py:class:`.CachedResponse` to a dict
+* Run json prconf converter to convert binary response body to base84
+* Run ``json.dumps()``
+
+For any optional libraries that aren't installed, the corresponding serializer will be a placeholder
+class that raises an ``ImportError`` at initialization time instead of at import time.
 """
 # flake8: noqa: F401
 from typing import Union
@@ -30,7 +48,6 @@ __all__ = [
     'pickle_serializer',
     'safe_pickle_serializer',
     'yaml_serializer',
-    'init_serializer',
     'utf8_encoder',
 ]
 
@@ -42,10 +59,3 @@ SERIALIZERS = {
 }
 
 SerializerType = Union[str, SerializerPipeline, Stage]
-
-
-def init_serializer(serializer: SerializerType = None):
-    """Initialize a serializer from a name or instance"""
-    if isinstance(serializer, str):
-        serializer = SERIALIZERS[serializer]
-    return serializer